+++ babel.config.js
@@ -3,11 +3,10 @@ const cjs = NODE_ENV === "test" || BABEL_ENV === "commonjs"
 const loose = true
 
 module.exports = {
-  presets: [["@babel/env", { loose, modules: false }]],
+  presets: [["@babel/env", { loose, modules: false }], "@babel/typescript"],
   plugins: [
     ["@babel/plugin-proposal-class-properties", { loose }],
     ["@babel/plugin-proposal-object-rest-spread", { loose }],
-    "@babel/plugin-transform-react-jsx",
     cjs && ["@babel/plugin-transform-modules-commonjs", { loose }],
     ["@babel/plugin-transform-runtime", { useESModules: !cjs }],
     "@babel/plugin-proposal-nullish-coalescing-operator",