+++ jest.config.js
@@ -1,5 +1,4 @@
 module.exports = {
   bail: true,
   verbose: true,
-  setupFilesAfterEnv: ["<rootDir>/testsSetup.js"],
 }