+++ rollup.config.js
@@ -2,14 +2,16 @@ import path from "path"
 
 import { split, compose, join, prepend, tail, map } from "ramda"
 import { toPascalCase, toKebabCase } from "ramda-extension"
-import autoExternal from "rollup-plugin-auto-external"
+import autoExternalPlugin from "rollup-plugin-auto-external"
 import babelPlugin from "rollup-plugin-babel"
 import cjsPlugin from "rollup-plugin-commonjs"
 import nodeResolvePlugin from "rollup-plugin-node-resolve"
-import replace from "rollup-plugin-replace"
+import replacePlugin from "rollup-plugin-replace"
 import { terser as terserPlugin } from "rollup-plugin-terser"
 
 const { LERNA_ROOT_PATH } = process.env
+
+const extensions = [".js", ".ts"]
 
 const plugins = {
   cjs: cjsPlugin({
@@ -23,9 +25,12 @@ const plugins = {
       warnings: false,
     },
   }),
-  nodeResolve: nodeResolvePlugin(),
+  nodeResolve: nodeResolvePlugin({
+    extensions,
+  }),
   babel: babelPlugin({
     cwd: LERNA_ROOT_PATH,
+    extensions,
     runtimeHelpers: true,
   }),
 }
@@ -36,11 +41,7 @@ const getFileName = compose(join("-"), prepend("afformative"), map(toKebabCase),
 
 const { LERNA_PACKAGE_NAME } = process.env
 const PACKAGE_ROOT_PATH = process.cwd()
-const INPUT_FILE = path.join(PACKAGE_ROOT_PATH, "src/index.js")
-
-const globals = {
-  react: "React",
-}
+const INPUT_FILE = path.join(PACKAGE_ROOT_PATH, "src/index.ts")
 
 const globalName = getGlobalName(LERNA_PACKAGE_NAME)
 const fileName = getFileName(LERNA_PACKAGE_NAME)
@@ -50,37 +51,35 @@ export default [
   {
     input: INPUT_FILE,
     output: {
-      file: path.join(PACKAGE_ROOT_PATH, "lib", `${fileName}.js`),
+      file: path.join(PACKAGE_ROOT_PATH, "dist", `${fileName}.cjs.js`),
       format: "cjs",
       indent: false,
     },
-    plugins: [autoExternal(), plugins.nodeResolve, plugins.babel, plugins.cjs],
+    plugins: [autoExternalPlugin(), plugins.nodeResolve, plugins.babel, plugins.cjs],
   },
 
   // NOTE: ES
   {
     input: INPUT_FILE,
     output: {
-      file: path.join(PACKAGE_ROOT_PATH, "es", `${fileName}.js`),
+      file: path.join(PACKAGE_ROOT_PATH, "dist", `${fileName}.esm.js`),
       format: "es",
       indent: false,
     },
-    plugins: [autoExternal(), plugins.nodeResolve, plugins.babel, plugins.cjs],
+    plugins: [autoExternalPlugin(), plugins.nodeResolve, plugins.babel, plugins.cjs],
   },
 
   // NOTE: UMD Development
   {
     input: INPUT_FILE,
-    external: Object.keys(globals),
     output: {
-      file: path.join(PACKAGE_ROOT_PATH, "dist", `${fileName}.js`),
+      file: path.join(PACKAGE_ROOT_PATH, "dist", `${fileName}.umd.js`),
       format: "umd",
       name: globalName,
       indent: false,
-      globals,
     },
     plugins: [
-      replace({ "process.env.NODE_ENV": JSON.stringify("development") }),
+      replacePlugin({ "process.env.NODE_ENV": JSON.stringify("development") }),
       plugins.nodeResolve,
       plugins.babel,
       plugins.cjs,
@@ -90,16 +89,14 @@ export default [
   // NOTE: UMD Production
   {
     input: INPUT_FILE,
-    external: Object.keys(globals),
     output: {
-      file: path.join(PACKAGE_ROOT_PATH, "dist", `${fileName}.min.js`),
+      file: path.join(PACKAGE_ROOT_PATH, "dist", `${fileName}.umd.min.js`),
       format: "umd",
       name: globalName,
       indent: false,
-      globals,
     },
     plugins: [
-      replace({ "process.env.NODE_ENV": JSON.stringify("production") }),
+      replacePlugin({ "process.env.NODE_ENV": JSON.stringify("production") }),
       plugins.nodeResolve,
       plugins.babel,
       plugins.cjs,